SGEA-main/requirements.txt
Django>=5.0,<6.0
djangorestframework
markdown
django-filter
